refactor: move compute stub handler to resource slice
